+++ amtest/src/Sass/Sass.scss
@@ -4,3 +4,4 @@
 @use '_CardsOfCharacters';
 @use '_NewRegister';
 @use '_Mixins';
+@use '_MediaQuerys';
