+++ amtest/src/Sass/_GlobalStyles.scss
@@ -1,4 +1,5 @@
 @use "./Mixins";
+@use "./MediaQuerys";
 @import url("https://fonts.googleapis.com/css2?family=Open+Sans:wght@400;600&display=swap");
 
 /* Text */
@@ -35,6 +36,13 @@ $noneHouse: 135deg, #7c7c7c 0%, #a8a8a8 100%;
 
 /*Generals*/
 
+html {
+  width: 100vw;
+  height: 100vh;
+  box-sizing: border-box;
+  font-family: 'Open Sans';
+}
+
 body {
   width: 100vw;
   height: 100vh;
@@ -66,7 +74,7 @@ span {
 
 .mainView__background {
   width: 100vw;
-  height: auto;
+  height:fit-content;
   background-size: cover;
   background-position: center center;
   position: absolute;
@@ -77,10 +85,16 @@ span {
 .mainCharacters {
   @include Mixins.flex(row, left, center);
   width: 80%;
-  height: 80%;
+  height: 65%;
   flex-wrap: wrap;
   margin-top: 5%;
   overflow-y: scroll;
+
+  @include MediaQuerys.phone {
+    width: 90%;
+    height: 70%;
+  }
+
 }
 
 .mainCharacters::-webkit-scrollbar {
