+++ amtest/src/Sass/_NewRegister.scss
@@ -1,5 +1,6 @@
 @use "../Sass/GlobalStyles";
 @use "./Mixins";
+@use "./MediaQuerys";
 
 .registerForm__hide {
   display: none;
@@ -32,15 +33,22 @@
     width: 90%;
     height: 15%;
 
-    &__close {
-      width: 3rem;
-      height: 3rem;
-      display: flex;
-    }
     &__title {
       color: GlobalStyles.$text-color-G;
       font-size: GlobalStyles.$Title-size;
       font-weight: GlobalStyles.$Bold;
+    }
+
+    &__close {
+      width: 7%;
+      height: 70%;
+      background-color: #ffff;
+      border: none;
+
+      &__icon {
+        width:100%;
+        height: 100%;
+      }
     }
   }
 
@@ -53,7 +61,6 @@
       @include Mixins.flex(column, flex-start, center);
       width: 100%;
       height: 50%;
-
 
       &__Section {
         @include Mixins.flex(row, space-between, center);
@@ -111,7 +118,6 @@
           width: 100%;
           height: 30%;
 
-
           &__radio {
             width: 2rem;
             margin-right: 1rem;
@@ -142,95 +148,16 @@
     &__button {
       @include Mixins.button;
     }
+
+    @include MediaQuerys.phone {
+      &__button {
+        font-size: 2vw;
+  
+      }
+    }
   }
+
+
+
+
 }
-// .register {
-//   width: 728px;
-//   background-color: #ffffff;
-//   border-radius: 15px;
-//   padding: 0 30px 30px;
-//   box-shadow: 0 4px 4px rgba(0, 0, 0, 0.25);
-
-//   &__header {
-//     display: flex;
-//     flex-direction: row;
-//     justify-content: space-between;
-//     align-items: center;
-//   }
-//   &__title {
-//     font-size: Globals.$title--form;
-//     font-weight: Globals.$bold;
-//     margin-bottom: 0px;
-//   }
-//   &__close {
-//     border: none;
-//     background-color: transparent;
-//   }
-
-//   &__img {
-//     width: 40px;
-//     height: 40px;
-//   }
-
-//   &__form {
-//     display: flex;
-//     flex-direction: row;
-//     flex-wrap: wrap;
-//     justify-content: space-around;
-
-//     .input {
-//       width: 310px;
-//       height: 74px;
-//       border-top-left-radius: 7px;
-//       border-top-right-radius: 7px;
-//       border: none;
-//       font-size: 16px;
-//       padding: 0 24px;
-//       background: rgba(#263238, 0.06);
-//       border-bottom: 2px solid rgba(#263238, 0.38);
-//       outline: none;
-//     }
-//     .form-group {
-//       display: flex;
-//       flex-direction: column;
-//       margin: 20px 0 0 0;
-//     }
-//   }
-// }
-
-// .label {
-//   margin: 10px 0;
-//   font-size: Globals.$rol;
-// }
-
-// .form-group__radio {
-//   display: flex;
-//   flex-direction: row;
-//   align-self: flex-start;
-//   width: 310px;
-// }
-
-// .form-group__label {
-//   margin: 0 0 0 25px;
-//   font-size: Globals.$form;
-//   font-weight: Globals.$regular;
-// }
-
-// .btn-save {
-//   margin: 32px 0 0 0;
-//   padding: 24px 144px;
-//   font-size: Globals.$button--filter;
-//   letter-spacing: 3px;
-//   background-color: Globals.$button--background--color;
-//   border: solid 3px Globals.$button--border--color;
-//   color: #ffffff;
-
-//   &:hover {
-//     background-color: rgba(107, 99, 181, 0.6);
-//     transition: 0.7s;
-//   }
-
-//   &:active {
-//     background-color: Globals.$button--border--color;
-//   }
-// }
