+++ amtest/src/Sass/_SortNav.scss
@@ -1,5 +1,6 @@
 @use "../Sass/GlobalStyles";
 @use "./Mixins";
+@use "./MediaQuerys";
 
 .nav {
   @include Mixins.flex(column, center, center);
@@ -34,4 +35,63 @@
       }
     }
   }
+
+  @include MediaQuerys.phone {
+    width: 100vw;
+    height: 25vh;
+    margin-top: 0vh;
+
+    &__img {
+      width: 40vw;
+      margin-bottom: .5rem;
+    }
+
+    &__text {
+      font-size: 3vh;
+
+    }
+
+    &__btns {
+      &__filter--active,
+      &__filter {
+        @include Mixins.buttonQueryphone;
+      }
+    }
+  }
+
+  @include MediaQuerys.desktop {
+    @include Mixins.flex(column, center, center);
+    width: 50vw;
+    margin-top: 10vh;
+  
+    &__img {
+      width: 20vw;
+      margin-bottom: 1rem;
+    }
+  
+    &__text {
+      color: GlobalStyles.$text-color-W;
+      font-size: GlobalStyles.$Title-size;
+      font-weight: GlobalStyles.$Bold;
+      margin-bottom: 1rem;
+    }
+  
+    &__btns {
+      @include Mixins.flex(row, center, center);
+  
+      &__filter--active,
+      &__filter {
+        @include Mixins.button;
+  
+        &--active {
+          background-color: GlobalStyles.$btn-color-active;
+        }
+  
+        &:hover {
+          background-color: GlobalStyles.$btn-color-hover;
+        }
+      }
+    }
+  }
+
 }
