--- amtest/src/css/NewRegister.css
@@ -1,242 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Open+Sans:wght@400;600&display=swap");
-/* Text */
-/*General Font Colors*/
-/*General Font Size*/
-/*General Font Weight*/
-/*Favorites Nav*/
-/* Buttons Colors*/
-/*Gradient Color by House*/
-/*Generals*/
-body {
-  width: 100vw;
-  height: 100vh;
-  margin: 0;
-  background-color: black;
-  font-family: "Open Sans";
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-p {
-  margin: 0;
-}
-
-button {
-  cursor: pointer;
-}
-
-span {
-  margin: 0;
-}
-
-/*MainVIEW*/
-.mainView {
-  display: flex;
-  flex-direction: column;
-  justify-content: none;
-  align-items: center;
-  width: 100vw;
-  height: 100vh;
-}
-
-.mainView__background {
-  width: 100vw;
-  height: auto;
-  background-size: cover;
-  background-position: center center;
-  position: absolute;
-  top: 0;
-  z-index: -1;
-}
-
-.mainCharacters {
-  display: flex;
-  flex-direction: row;
-  justify-content: left;
-  align-items: center;
-  width: 80%;
-  height: 80%;
-  flex-wrap: wrap;
-  margin-top: 5%;
-  overflow-y: scroll;
-}
-
-.mainCharacters::-webkit-scrollbar {
-  width: 0.1vw;
-}
-
-.mainCharacters::-webkit-scrollbar-track {
-  background-color: rgba(0, 0, 0, 0);
-}
-
-.mainCharacters::-webkit-scrollbar-thumb {
-  background-color: rgba(0, 0, 0, 0);
-  height: 0%;
-}
-
-.registerForm__hide {
-  display: none;
-}
-
-.registerForm__isActive {
-  display: block;
-  width: 100vw;
-  height: 100vh;
-  position: fixed;
-  z-index: 4;
-  left: 0;
-  top: 0;
-  overflow: auto;
-  background-color: rgba(0, 0, 0, 0.4);
-}
-
-.FormRegister {
-  width: 80vw;
-  height: 50vh;
-  background-color: white;
-  position: absolute;
-  top: 25vh;
-  left: 10vw;
-  z-index: 3;
-  border-radius: 15px;
-  display: flex;
-  flex-direction: column;
-  justify-content: none;
-  align-items: center;
-}
-.FormRegister__Header {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  width: 90%;
-  height: 15%;
-}
-.FormRegister__Header__close {
-  width: 3rem;
-  height: 3rem;
-  display: flex;
-}
-.FormRegister__Header__title {
-  color: #081f32;
-  font-size: 3vw;
-  font-weight: 600;
-}
-.FormRegister__Form {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: center;
-  width: 90%;
-  height: 80%;
-}
-.FormRegister__Form__inputs {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: center;
-  width: 100%;
-  height: 50%;
-}
-.FormRegister__Form__inputs__Section {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  height: 50%;
-}
-.FormRegister__Form__inputs__Section__group {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: none;
-  width: 47%;
-  height: 85%;
-}
-.FormRegister__Form__inputs__Section__group__label {
-  color: #081f32;
-  font-size: 1.9vw;
-  font-weight: 400;
-  margin-bottom: 0.5rem;
-}
-.FormRegister__Form__inputs__Section__group__input {
-  height: 90%;
-  color: #081f32;
-  font-size: 1.9vw;
-  font-weight: 400;
-  padding: 0 25px;
-  background: rgba(38, 50, 56, 0.06);
-  border: none;
-  border-bottom: 3px solid rgba(38, 50, 56, 0.38);
-  border-top-left-radius: 7px;
-  border-top-right-radius: 7px;
-  outline: none;
-}
-.FormRegister__Form__checks {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: flex-start;
-  width: 100%;
-  height: 20%;
-}
-.FormRegister__Form__checks__Section2 {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: baseline;
-  width: 47%;
-  height: 100%;
-}
-.FormRegister__Form__checks__Section2__label {
-  color: #081f32;
-  font-size: 1.9vw;
-  font-weight: 400;
-  margin: 0.5rem 0;
-}
-.FormRegister__Form__checks__Section2__group {
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-  width: 100%;
-  height: 30%;
-}
-.FormRegister__Form__checks__Section2__group__radio {
-  width: 2rem;
-  margin-right: 1rem;
-}
-.FormRegister__Form__checks__Section2__group__radio__label {
-  color: #081f32;
-  font-size: 1.9vw;
-  font-weight: 400;
-  margin-right: 20%;
-}
-.FormRegister__Form__inputFile {
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-  width: 100%;
-  height: 10%;
-}
-.FormRegister__Form__inputFile__label {
-  color: #081f32;
-  font-size: 1.9vw;
-  font-weight: 400;
-  margin-right: 2rem;
-}
-.FormRegister__Form__button {
-  width: 23vw;
-  padding: 2vh;
-  border: 4px solid #6B63B5;
-  color: white;
-  font-size: 1vw;
-  font-weight: 600;
-  letter-spacing: 3px;
-  margin: 1rem;
-  background-color: rgba(0, 0, 0, 0.6);
-}
-
-/*# sourceMappingURL=NewRegister.css.map */
